@@ -138,13 +134,14 @@
     "@types/react-select": "^3.1.2",
     "@types/request": "^2.48.8",
     "@types/styled-components": "^5.1.26",
+    "@types/url-parse": "^1.4.8",
     "@types/wavesurfer.js": "^3.3.2",
     "concurrently": "^6.5.1",
-    "craco-plugin-react-hot-reload": "^0.1.0",
-    "electron": "^11.3.0",
-    "electron-builder": "22.10.5",
+    "cross-env": "^7.0.3",
+    "electron": "^22.0.0",
+    "electron-builder": "23.6.0",
     "electron-debug": "^3.2.0",
-    "electron-devtools-installer": "^3.1.1",
+    "electron-devtools-installer": "^3.2.0",
     "electron-react-devtools": "^0.5.3",
     "env-cmd": "^10.1.0",
     "md5-file": "^5.0.0",
