@@ -1,16 +1,11 @@
-import { useGlobal } from 'reactn';
 import { ISharedStrings, IState, Group } from '../model';
 import { connect } from 'react-redux';
 import { QueryBuilder } from '@orbit/data';
 import { withData } from 'react-orbitjs';
 import { Avatar, AvatarProps, styled } from '@mui/material';
 import { makeAbbr } from '../utils';
-import { dataPath, PathType } from '../utils/dataPath';
-import { remoteId } from '../crud';
-import { isElectron } from '../api-variable';
 import localStrings from '../selector/localize';
-const os = require('os');
-const fs = require('fs');
+import { useAvatarSource } from '../crud';
 
 // see: https://mui.com/material-ui/customization/how-to-customize/
 interface StyledAvatarProps extends AvatarProps {
@@ -45,23 +40,10 @@ interface IProps extends IStateProps, IRecordProps {
 
 export function GroupAvatar(props: IProps) {
   const { groupRec, small } = props;
-  const [memory] = useGlobal('memory');
+  const source = useAvatarSource(groupRec.attributes.name, groupRec);
 
-  var src = dataPath(groupRec.attributes.name, PathType.AVATARS, {
-    localname:
-      remoteId('group', groupRec.id, memory.keyMap) +
-      groupRec.attributes.name +
-      '.png',
-  });
-  if (src && isElectron && !src.startsWith('http')) {
-    if (fs.existsSync(src)) {
-      const url =
-        os.platform() === 'win32' ? new URL(src).toString().slice(8) : src;
-      src = `transcribe-safe://${url}`;
-    } else src = '';
-  }
-  return src ? (
-    <StyledAvatar alt={groupRec.attributes.name} src={src} small={small} />
+  return source ? (
+    <StyledAvatar alt={groupRec.attributes.name} src={source} small={small} />
   ) : groupRec.attributes && groupRec.attributes.name !== '' ? (
     <StyledAvatar small={small}>
       {makeAbbr(groupRec.attributes.name)}
