@@ -1,15 +1,12 @@
-import fs from 'fs';
-import path from 'path';
-import { isElectron } from '../api-variable';
-const ipc = isElectron ? require('electron').ipcRenderer : null;
-const os = require('os');
+import path from 'path-browserify';
+const ipc = (window as any)?.electron;
 
 export const audPrefsName = async () => {
-  if (os.platform() === 'win32') {
-    const appData = await ipc?.invoke('appData');
+  if (await ipc?.isWindows()) {
+    const appData = await ipc?.appData();
     return path.join(appData, 'audacity', 'audacity.cfg');
   } else {
-    const home = await ipc?.invoke('home');
+    const home = await ipc?.home();
     const snapName = path.join(
       home,
       'snap',
@@ -18,15 +15,15 @@ export const audPrefsName = async () => {
       '.audacity-data',
       'audacity.cfg'
     );
-    if (fs.existsSync(snapName)) return snapName;
+    if (await ipc?.exists(snapName)) return snapName;
     // const debName = path.join(home, '.audacity-data', 'audacity.cfg');
     // if (fs.existsSync(debName)) return debName;
   }
 };
 
 const getAllPref = async (prefs: string) => {
-  if (!fs.existsSync(prefs)) return false;
-  return fs.readFileSync(prefs, 'utf-8');
+  if (!(await ipc?.exists(prefs))) return false;
+  return await ipc?.read(prefs, 'utf-8');
 };
 
 const getAudPrefContent = async (inPrefs?: string) => {
@@ -55,13 +52,13 @@ const getIoMatch = (content: string) => {
   );
 };
 
-const setFolder = (kind: string, target: string) => {
-  const ln = os.platform() === 'win32' ? '\r\n' : '\n';
+const setFolder = async (kind: string, target: string) => {
+  const ln = (await ipc?.isWindows()) ? '\r\n' : '\n';
   return `[Directories/${kind}]${ln}Default=${target}${ln}LastUsed=${target}${ln}`;
 };
 
 export const resetAudacityPref = (prefs: string, data: string) => {
-  fs.writeFileSync(prefs, data);
+  ipc?.write(prefs, data);
 };
 
 const changeValueTo1 = (data: string, m: RegExpExecArray) => {
@@ -71,9 +68,9 @@ const changeValueTo1 = (data: string, m: RegExpExecArray) => {
 
 export const setAudacityPref = async (fullName: string) => {
   let folder = path.dirname(fullName);
-  fs.mkdirSync(folder, { recursive: true });
+  await ipc?.createFolder(folder);
   let ioFolder = folder.replace('aup3', 'io');
-  fs.mkdirSync(ioFolder, { recursive: true });
+  await ipc?.createFolder(ioFolder);
   if (path.sep === '\\') {
     folder = folder.replace(/\\/g, `\\\\`);
     ioFolder = ioFolder.replace(/\\/g, `\\\\`);
