@@ -1,6 +1,8 @@
 const { BrowserWindow, Menu, app } = require('electron');
 const authService = require('./auth-service');
 const createAppWindow = require('./app-process');
+const path = require('path');
+const isDev = require('electron-is-dev');
 
 let win = null;
 
@@ -10,9 +12,17 @@ function createAuthWindow(hasUsed, email) {
   win = new BrowserWindow({
     width: 1000,
     height: 780,
+    icon: path.join(__dirname, 'favicon.ico'),
     webPreferences: {
+      devTools: isDev,
       nodeIntegration: false,
+      nodeIntegrationInWorker: false,
+      nodeIntegrationInSubFrames: false,
       enableRemoteModule: false,
+      contextIsolation: true,
+      webSecurity: false,
+      spellcheck: true,
+      preload: path.join(__dirname, 'preload.js'),
     },
   });
 
@@ -31,7 +41,7 @@ function createAuthWindow(hasUsed, email) {
             return workOffline();
           },
         },
-        // { role: 'toggleDevTools' },
+        ...(isDev ? [{ role: 'toggleDevTools' }] : []),
         {
           label: 'Exit',
           click() {
@@ -72,6 +82,10 @@ function createAuthWindow(hasUsed, email) {
   win.on('authenticated', () => {
     destroyAuthWin();
   });
+
+  if (isDev) {
+    win.webContents.openDevTools({ mode: 'detach' });
+  }
 
   win.on('closed', () => {
     win = null;
