@@ -10,12 +10,11 @@ import DownloadIcon from '@mui/icons-material/CloudDownload';
 import AddIcon from '@mui/icons-material/Add';
 import RemoveIcon from '@mui/icons-material/Remove';
 import { StyledMenu, StyledMenuItem } from '../control';
-import path from 'path';
+import path from 'path-browserify';
 import { isElectron, API_CONFIG } from '../api-variable';
 import {
   launch,
   launchCmd,
-  downloadFile,
   dataPath,
   PathType,
   execFolder,
@@ -26,7 +25,7 @@ import AboutDialog from './AboutDialog';
 import { usePlan, remoteIdGuid } from '../crud';
 import { mainSelector } from '../selector';
 import { shallowEqual, useSelector } from 'react-redux';
-const os = require('os');
+const ipc = (window as any)?.electron;
 
 interface IProps {
   online: boolean;
@@ -81,23 +80,26 @@ export function HelpMenu(props: IProps) {
     return 'en';
   };
 
-  const handleHelp = (topic?: string) => () => {
+  const handleHelp = (topic?: string) => async () => {
+    const chmHelp = API_CONFIG.chmHelp;
+    const helpUrl = API_CONFIG.help;
     const topicS = topic || '';
     const topicWin = topic && decodeURIComponent(topic.slice(3));
     if (isElectron) {
+      const folder = await execFolder();
       // see https://stackoverflow.com/questions/22300244/open-a-chm-file-to-a-specific-topic
-      if (os.platform() === 'win32' && topicWin && !online) {
+      if ((await ipc?.isWindows()) && topicWin && !online) {
         const target = `C:\\Windows\\hh.exe ${path.join(
-          execFolder(),
-          API_CONFIG.chmHelp
+          folder,
+          chmHelp
         )}::${topicWin}`;
         launchCmd(target);
       } else if (topic && !online) {
-        launchCmd(`xchm -c 1 ${path.join(execFolder(), API_CONFIG.chmHelp)}`);
+        launchCmd(`xchm -c 1 ${path.join(folder, chmHelp)}`);
       } else {
         const target = !online
-          ? path.join(execFolder(), API_CONFIG.chmHelp)
-          : API_CONFIG.help + '/' + helpLanguage() + indexName + topicS;
+          ? path.join(folder, chmHelp)
+          : helpUrl + '/' + helpLanguage() + indexName + topicS;
         launch(target, online);
       }
     } else if (helpRef.current) {
@@ -111,10 +113,11 @@ export function HelpMenu(props: IProps) {
     const loc = inOffline !== undefined ? inOffline : offline;
     const urlObj = new URL(url);
     const name = urlObj.pathname.split('/').pop() || '';
+    const folder = await execFolder();
     const localPath = loc
-      ? path.join(execFolder(), 'help', name)
+      ? path.join(folder, 'help', name)
       : dataPath(name, PathType.ZIP);
-    if (!loc) await downloadFile({ url, localPath });
+    if (!loc) await ipc?.downloadFile(url, localPath);
     launch(localPath, false);
     setAnchorEl(null);
     if (action) action('Download');
