@@ -20,7 +20,6 @@ import { offlineProjectUpdateFilesDownloaded, useProjectExport } from '../crud';
 import {
   currentDateTime,
   dataPath,
-  downloadFile,
   logError,
   PathType,
   Severity,
@@ -28,6 +27,8 @@ import {
 import AdmZip from 'adm-zip';
 import { Operation } from '@orbit/data';
 import IndexedDBSource from '@orbit/indexeddb';
+const ipc = (window as any)?.electron;
+
 enum Steps {
   Prepare,
   Download,
@@ -143,11 +144,12 @@ export const ProjectDownload = (
   React.useEffect(() => {
     if (progress === Steps.Download) {
       const localPath = dataPath(exportName, PathType.ZIP);
-      downloadFile({ url: exportUrl, localPath })
+      ipc
+        ?.downloadFile(exportUrl, localPath)
         .then(() => {
           setProgress(Steps.Import);
         })
-        .catch((ex) => logError(Severity.error, errorReporter, ex))
+        .catch((ex: any) => logError(Severity.error, errorReporter, ex))
         .finally(() => {
           URL.revokeObjectURL(exportUrl);
         });
@@ -161,17 +163,20 @@ export const ProjectDownload = (
 
   React.useEffect(() => {
     if (progress === Steps.Import) {
-      const localPath = dataPath(exportName, PathType.ZIP);
-      const zip = new AdmZip(localPath);
-      zip.extractAllTo(dataPath(), true);
-      offlineProjectUpdateFilesDownloaded(
-        projectIds[currentStep],
-        offlineUpdates,
-        memory,
-        currentDateTime()
-      );
-      setProgress(Steps.Prepare);
-      setCurrentStep(currentStep + 1);
+      (async () => {
+        const localPath = dataPath(exportName, PathType.ZIP);
+        const zip = (await ipc?.zipOpen(localPath)) as AdmZip;
+        await ipc?.zipExtract(zip, dataPath(), true);
+        await ipc?.zipClose(zip);
+        offlineProjectUpdateFilesDownloaded(
+          projectIds[currentStep],
+          offlineUpdates,
+          memory,
+          currentDateTime()
+        );
+        setProgress(Steps.Prepare);
+        setCurrentStep(currentStep + 1);
+      })();
     } else if (progress === Steps.Error) {
       setProgress(Steps.Prepare);
       setCurrentStep(currentStep + 1);
