@@ -7,7 +7,7 @@ import { remoteIdGuid, remoteId } from '../crud';
 import { dataPath, PathType } from '../utils/dataPath';
 import { MediaFile } from '../model';
 import { infoMsg, logError, Severity } from '../utils';
-const os = require('os');
+const ipc = (window as any)?.electron;
 // See: https://www.smashingmagazine.com/2020/07/custom-react-hook-fetch-cache-data/
 
 export enum MediaSt {
@@ -88,9 +88,9 @@ export const useFetchMediaUrl = (reporter?: any) => {
     return isNaN(Number(id)) ? remoteId('mediafile', id, memory.keyMap) : id;
   };
 
-  const safeURL = (path: string) => {
+  const safeURL = async (path: string) => {
     if (!path.startsWith('http')) {
-      const start = os.platform() === 'win32' ? 8 : 7;
+      const start = (await ipc?.isWindows()) ? 8 : 7;
       const url = new URL(`file://${path}`).toString().slice(start);
       return `transcribe-safe://${url}`;
     }
@@ -109,7 +109,7 @@ export const useFetchMediaUrl = (reporter?: any) => {
       return false;
     };
 
-    const fetchData = () => {
+    const fetchData = async () => {
       if (isElectron) {
         try {
           if (cancelled()) return;
@@ -125,7 +125,7 @@ export const useFetchMediaUrl = (reporter?: any) => {
             const path = dataPath(audioUrl, PathType.MEDIA);
             if (!path.startsWith('http')) {
               if (cancelled()) return;
-              dispatch({ payload: safeURL(path), type: MediaSt.FETCHED });
+              dispatch({ payload: await safeURL(path), type: MediaSt.FETCHED });
               return;
             } else if (!accessToken) {
               dispatch({
