@@ -1,7 +1,8 @@
 import Axios from 'axios';
 import { FETCH_LOCALIZATION, SET_LANGUAGE } from './types';
 import { appPath } from '../../utils';
-import { stringsName } from './exported-strings-name.json';
+import exportStrings from './exported-strings-name.json';
+const { stringsName } = exportStrings;
 
 export const fetchLocalization = () => (dispatch: any) => {
   Axios.get(appPath() + '/localization/' + stringsName).then((strings) => {
