@@ -19,8 +19,8 @@ import {
 } from '../../utils';
 import moment from 'moment';
 import _ from 'lodash';
-var fs = require('fs');
-var path = require('path');
+const ipc = (window as any)?.electron;
+var path = require('path-browserify');
 
 export const uploadFiles = (files: File[]) => (dispatch: any) => {
   dispatch({
@@ -38,28 +38,35 @@ const nextVersion = (fileName: string) => {
   return `${name}.ver02.${ext}`;
 };
 
-export const writeFileLocal = (file: File, remoteName?: string) => {
+let writeName = ''; // used for message if copy fails
+
+export const writeFileLocal = async (file: File, remoteName?: string) => {
   var local = { localname: '' };
+  const filePath = (file as any)?.path || '';
   dataPath(
-    remoteName ? remoteName : `http://${file.path}`,
+    remoteName ? remoteName : `http://${filePath}`,
     PathType.MEDIA,
     local
   );
-  var fullName = local.localname;
-  if (!remoteName && file.path === '') fullName += path.sep + file.name;
-  createPathFolder(fullName);
-  while (fs.existsSync(fullName)) {
-    fullName = nextVersion(fullName);
+  writeName = local.localname;
+  if (!remoteName && filePath === '') writeName += path.sep + file.name;
+  await createPathFolder(writeName);
+  while (await ipc?.exists(writeName)) {
+    writeName = nextVersion(writeName);
   }
-  const reader = new FileReader();
-  reader.onload = (evt) => {
-    fs.writeFileSync(fullName, evt?.target?.result, {
-      encoding: 'binary',
-      flag: 'wx', //write - fail if file exists
-    });
-  };
-  reader.readAsBinaryString(file);
-  return path.join(PathType.MEDIA, fullName.split(path.sep).pop());
+  if (filePath) {
+    await ipc?.copyFile(filePath, writeName);
+  } else {
+    const reader = new FileReader();
+    reader.onload = (evt) => {
+      ipc?.write(writeName, evt?.target?.result, {
+        encoding: 'binary',
+        flag: 'wx', //write - fail if file exists
+      });
+    };
+    reader.readAsBinaryString(file);
+  }
+  return path.join(PathType.MEDIA, writeName.split(path.sep).pop());
 };
 const uploadFile = (
   data: any,
@@ -230,18 +237,21 @@ export const nextUpload =
         Authorization: 'Bearer ' + token,
       },
     })
-      .then((response) => {
+      .then(async (response) => {
         dispatch({ payload: n, type: UPLOAD_ITEM_CREATED });
         var json = fromVnd(response.data);
         uploadFile(json, files[n], errorReporter, token, completeCB);
         if (isElectron) {
           try {
-            writeFileLocal(files[n], response.data.audioUrl);
+            await writeFileLocal(
+              files[n],
+              response.data.data.attributes['audio-url']
+            );
           } catch (err) {
             logError(
               Severity.error,
               errorReporter,
-              `failed writing ${files[n]}`
+              `failed copying ${(files[n] as any).path} to ${writeName}`
             );
           }
         }
