@@ -1,9 +1,11 @@
-import path from 'path';
+import path from 'path-browserify';
+const ipc = (window as any)?.electron;
 
-export const execFolder = () =>
-  path.dirname(
-    (process as any).helperExecPath.replace(
-      path.join('node_modules', 'electron', 'dist'),
-      path.join('dist', 'win-unpacked')
-    )
-  );
+export const execFolder = async () => {
+  const folder = await ipc?.execPath();
+  const fromStr = path.join('node_modules', 'electron', 'dist');
+  const toStr = path.join('dist', 'win-unpacked');
+  const replaced = folder.replace(fromStr, toStr);
+  const result = path.dirname(replaced);
+  return result;
+};
