@@ -1,14 +1,7 @@
-const isElectron = process.env.REACT_APP_MODE === 'electron';
-var fs = isElectron ? require('fs-extra') : null;
-const convert = require('xml-js');
+const ipc = (window as any)?.electron;
 
-export const fileJson = (settings: string) => {
-  if (fs.existsSync(settings)) {
-    const data = fs.readFileSync(settings, 'utf-8');
-    const jsonStr = convert.xml2json(data, { compact: true, spaces: 2 });
-    return JSON.parse(jsonStr);
-  }
-  return null;
+export const fileJson = async (settings: string) => {
+  return JSON.parse(await ipc?.fileJson(settings));
 };
 
 export default fileJson;
