@@ -1,11 +1,11 @@
 import { getRegVal } from '.';
 import { getWhereis } from './getWhereis';
-const os = require('os');
+const ipc = (window as any)?.electron;
 
 const key = 'HKCR\\Audacity.Project\\shell\\open\\command';
 
 export const hasAudacity = async () => {
-  if (os.platform() === 'win32') {
+  if (await ipc?.isWindows()) {
     return Boolean(await getRegVal(key, ''));
   } else {
     return (await getWhereis('audacity')).length > 0;
