@@ -1,10 +1,10 @@
 import { getRegVal } from '.';
 import { getWhereis } from './getWhereis';
-const os = require('os');
+const ipc = (window as any)?.electron;
 
 const key = 'HKCR\\Python.CompiledFile\\shell\\open\\command';
 export const hasPython = async () => {
-  if (os.platform() === 'win32') {
+  if (await ipc?.isWindows()) {
     return Boolean(await getRegVal(key, ''));
   } else {
     return (await getWhereis('python')).length > 0;
