@@ -1,43 +1,38 @@
-import { isElectron } from '../api-variable';
-const ipc = isElectron ? require('electron').ipcRenderer : null;
-const fs = isElectron ? require('fs-extra') : null;
-const path = require('path');
+const ipc = (window as any)?.electron;
+const path = require('path-browserify');
 
-const noop = {} as any;
-const { shell } = isElectron ? require('electron') : { shell: noop };
-const execa = isElectron ? require('execa') : noop;
-const os = require('os');
-
-export const launch = (target: string, online: boolean) => {
+export const launch = async (target: string, online: boolean) => {
   if (/\.pdf$/i.test(target)) target = target.slice(18);
-  if (online) shell.openExternal(target);
-  else if (os.platform() === 'win32') shell.openPath('file:///' + target);
+  if (online) ipc?.openExternal(target);
+  else if (await ipc?.isWindows()) ipc?.openPath('file:///' + target);
   else {
     console.log(`launching ${target}`);
     const cmd = /\.sh/i.test(target) ? '' : 'xdg-open ';
-    execa.command(`${cmd}${target}`, {
+    ipc?.exeCmd(`${cmd}${target}`, {
       env: { ...{ ...process }.env, DISPLAY: ':0' },
     });
   }
 };
 
 export const launchCmd = async (target: string) => {
-  const temp = await ipc?.invoke('temp');
+  const temp = await ipc?.temp();
   if (!temp) throw new Error('Unable to find temp directory.'); //this is app.getPath('temp')
-  if (os.platform() === 'win32') {
+  if (await ipc?.isWindows()) {
     const tempName = path.join(temp, 'transcriber-cmd.ps1');
-    fs.writeFileSync(tempName, target, { encoding: 'utf-8' });
-    execa(`powershell`, [tempName]).finally(() => {
-      fs.unlinkSync(tempName);
+    await ipc?.write(tempName, target);
+    ipc?.exec(`powershell`, [tempName]).finally(async () => {
+      await ipc?.delete(tempName);
     });
   } else {
     const tempName = path.join(temp, 'transcriber-cmd.sh');
-    fs.writeFileSync(tempName, target, { encoding: 'utf-8' });
-    execa(`sh`, [tempName], {
-      env: { ...{ ...process }.env, DISPLAY: ':0' },
-    }).finally(() => {
-      fs.unlinkSync(tempName);
-    });
+    ipc?.write(tempName, target);
+    ipc
+      ?.exec(`sh`, [tempName], {
+        env: { ...{ ...process }.env, DISPLAY: ':0' },
+      })
+      .finally(async () => {
+        await ipc?.unlinkSync(tempName);
+      });
   }
 };
 
